LayoutTests/imported/w3c/web-platform-tests/html/webappapis/timers/negative-settimeout-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_unreached: Reached unreachable code

PASS Negative timeout in setTimeout 
